refactor(pw): hide internals of struct pipewire
